requirements.txt
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==2.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.27.1
urllib3==1.26.20
Werkzeug==2.0.3